+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinExternalPay/Merchant/CgibinExternalPaySetMerchantUseScopeRequest.cs
@@ -23,7 +23,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("partyid")]
                 [System.Text.Json.Serialization.JsonPropertyName("partyid")]
-                public IList<int>? DepartmentIdList { get; set; }
+                public IList<long>? DepartmentIdList { get; set; }
 
                 /// <summary>
                 /// 获取或设置标签 ID 列表。
