+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/Living/CgibinSchoolLivingGetUnwatchStatisticsResponse.cs
@@ -32,7 +32,7 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("partyids")]
                         [System.Text.Json.Serialization.JsonPropertyName("partyids")]
-                        public int[] DepartmentIdList { get; set; } = default!;
+                        public long[] DepartmentIdList { get; set; } = default!;
                     }
                 }
