+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/User/CgibinSchoolUserGetResponse.cs
@@ -71,7 +71,7 @@
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("department")]
                 [System.Text.Json.Serialization.JsonPropertyName("department")]
-                public int[] DepartmentIdList { get; set; } = default!;
+                public long[] DepartmentIdList { get; set; } = default!;
 
                 /// <summary>
                 /// 获取或设置学生的家长列表。
