+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinUser/CgibinUserCreateRequest.cs
@@ -263,7 +263,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("department")]
         [System.Text.Json.Serialization.JsonPropertyName("department")]
-        public IList<int> DepartmentIdList { get; set; } = new List<int>();
+        public IList<long> DepartmentIdList { get; set; } = new List<long>();
 
         /// <summary>
         /// 获取或设置部门次序列表。
@@ -284,7 +284,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("main_department")]
         [System.Text.Json.Serialization.JsonPropertyName("main_department")]
-        public int? MainDepartmentId { get; set; }
+        public long? MainDepartmentId { get; set; }
 
         /// <summary>
         /// 获取或设置直属上级成员账号列表。
