+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinUser/CgibinUserSimpleListResponse.cs
@@ -35,7 +35,7 @@
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("department")]
                 [System.Text.Json.Serialization.JsonPropertyName("department")]
-                public int[] DepartmentIdList { get; set; } = default!;
+                public long[] DepartmentIdList { get; set; } = default!;
             }
         }
